+++ packages/flutter_chat/lib/src/flutter_chat_userstory.dart
@@ -14,9 +14,11 @@ List<GoRoute> getChatStoryRoutes(
       GoRoute(
         path: ChatUserStoryRoutes.chatScreen,
         pageBuilder: (context, state) {
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var chatScreen = ChatScreen(
             unreadMessageTextStyle: configuration.unreadMessageTextStyle,
-            service: configuration.chatService,
+            service: service,
             options: configuration.chatOptionsBuilder(context),
             onNoChats: () async =>
                 context.push(ChatUserStoryRoutes.newChatScreen),
@@ -34,7 +36,8 @@ List<GoRoute> getChatStoryRoutes(
                 configuration.onDeleteChat?.call(context, chat) ??
                 configuration.chatService.chatOverviewService.deleteChat(chat),
             deleteChatDialog: configuration.deleteChatDialog,
-            translations: configuration.translations,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
           );
           return buildScreenWithoutTransition(
             context: context,
@@ -53,6 +56,8 @@ List<GoRoute> getChatStoryRoutes(
         path: ChatUserStoryRoutes.chatDetailScreen,
         pageBuilder: (context, state) {
           var chatId = state.pathParameters['id'];
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var chatDetailScreen = ChatDetailScreen(
             chatTitleBuilder: configuration.chatTitleBuilder,
             usernameBuilder: configuration.usernameBuilder,
@@ -60,8 +65,9 @@ List<GoRoute> getChatStoryRoutes(
             iconDisabledColor: configuration.iconDisabledColor,
             pageSize: configuration.messagePageSize,
             options: configuration.chatOptionsBuilder(context),
-            translations: configuration.translations,
-            service: configuration.chatService,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
+            service: service,
             chatId: chatId!,
             textfieldBottomPadding: configuration.textfieldBottomPadding ?? 0,
             onPressUserProfile: (userId) async {
@@ -120,10 +126,13 @@ List<GoRoute> getChatStoryRoutes(
       GoRoute(
         path: ChatUserStoryRoutes.newChatScreen,
         pageBuilder: (context, state) {
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var newChatScreen = NewChatScreen(
             options: configuration.chatOptionsBuilder(context),
-            translations: configuration.translations,
-            service: configuration.chatService,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
+            service: service,
             onPressCreateChat: (user) async {
               configuration.onPressCreateChat?.call(user);
               if (configuration.onPressCreateChat != null) return;
@@ -163,10 +172,13 @@ List<GoRoute> getChatStoryRoutes(
       GoRoute(
         path: ChatUserStoryRoutes.newGroupChatScreen,
         pageBuilder: (context, state) {
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var newGroupChatScreen = NewGroupChatScreen(
             options: configuration.chatOptionsBuilder(context),
-            translations: configuration.translations,
-            service: configuration.chatService,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
+            service: service,
             onPressGroupChatOverview: (users) async => context.push(
               ChatUserStoryRoutes.newGroupChatOverviewScreen,
               extra: users,
@@ -188,11 +200,14 @@ List<GoRoute> getChatStoryRoutes(
       GoRoute(
         path: ChatUserStoryRoutes.newGroupChatOverviewScreen,
         pageBuilder: (context, state) {
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var users = state.extra! as List<ChatUserModel>;
           var newGroupChatOverviewScreen = NewGroupChatOverviewScreen(
             options: configuration.chatOptionsBuilder(context),
-            translations: configuration.translations,
-            service: configuration.chatService,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
+            service: service,
             users: users,
             onPressCompleteGroupChatCreation: (users, groupChatName) async {
               configuration.onPressCompleteGroupChatCreation
@@ -232,9 +247,12 @@ List<GoRoute> getChatStoryRoutes(
           var chatId = state.pathParameters['id'];
           var userId = state.pathParameters['userId'];
           var id = userId == 'null' ? null : userId;
+          var service = configuration.chatServiceBuilder?.call(context) ??
+              configuration.chatService;
           var profileScreen = ChatProfileScreen(
-            translations: configuration.translations,
-            chatService: configuration.chatService,
+            translations: configuration.translationsBuilder?.call(context) ??
+                configuration.translations,
+            chatService: service,
             chatId: chatId!,
             userId: id,
             onTapUser: (user) async {